refactor: add setup and teardown for dummy files in tests
