+++ server/__tests__/e2e/files.projects.test.ts
@@ -1,15 +1,9 @@
 import fetch from 'node-fetch';
-import { createNewUserAndLogin } from '../utils';
-import { FileHandle, open, mkdtemp, rm } from 'node:fs/promises';
-import { join } from 'node:path';
-import { tmpdir } from 'node:os';
-import FormData from 'form-data';
-import { createReadStream } from 'node:fs';
-import { S3Client, DeleteObjectsCommand } from '@aws-sdk/client-s3';
-
-const bucketName = process.env.AWS_S3_BUCKET;
-const region = process.env.AWS_BUCKET_REGION;
-const s3 = new S3Client({ region });
+import {
+  createNewUserAndLogin,
+  dummyFilesSetup,
+  dummyFilesTeardown,
+} from '../utils';
 
 describe('files endpoint, projects auth...', () => {
   it('allows a user to upload project media to a project they are an admin for...', async () => {
@@ -40,50 +34,16 @@ describe('files endpoint, projects auth...', () => {
     const key = `projects-test/${projectId}`;
     const fname = `test-${projectId}.txt`;
 
-    let dir: undefined | string = undefined;
-    let fd: undefined | FileHandle = undefined;
     try {
-      const path = join(tmpdir(), 'projects-');
-      dir = await mkdtemp(path);
-      const fpath = join(dir, fname);
-      fd = await open(fpath, 'w');
-      await fd.write(`helloworld, for ${projectId}.`);
-      await fd.sync();
-
-      const form = new FormData();
-      form.append('image', createReadStream(fpath));
-      form.append('filePath', key);
-
-      authHeaders.delete('content-type');
-      const resp = await fetch('http://localhost:5000/files', {
-        method: 'POST',
-        headers: authHeaders,
-        body: form,
-      });
+      const { resp } = await dummyFilesSetup(
+        key,
+        fname,
+        projectId,
+        authHeaders,
+      );
       expect(resp.status).toBe(200);
     } finally {
-      if (fd) {
-        await fd?.close();
-      }
-      if (dir) {
-        await rm(dir, { recursive: true, force: true });
-      }
-      const cmd = new DeleteObjectsCommand({
-        Bucket: bucketName,
-        Delete: {
-          Objects: [{ Key: `${key}/${fname}` }, { Key: `${key}` }],
-        },
-      });
-      const response = await s3.send(cmd);
-      if (response['$metadata'].httpStatusCode !== 200) {
-        console.dir(
-          {
-            response,
-            warning: 'objects possibly not removed rom s3 testing bucket',
-          },
-          { depth: null },
-        );
-      }
+      await dummyFilesTeardown(key, fname);
     }
   });
   it('disallows a user from uploading project media to a project they are not admin for...', async () => {
@@ -117,34 +77,7 @@ describe('files endpoint, projects auth...', () => {
     const key = `projects-test/${projectId}`;
     const fname = `test-${projectId}.txt`;
 
-    let dir: undefined | string = undefined;
-    let fd: undefined | FileHandle = undefined;
-    try {
-      const path = join(tmpdir(), 'projects-');
-      dir = await mkdtemp(path);
-      const fpath = join(dir, fname);
-      fd = await open(fpath, 'w');
-      await fd.write(`helloworld, for ${projectId}.`);
-      await fd.sync();
-
-      const form = new FormData();
-      form.append('image', createReadStream(fpath));
-      form.append('filePath', key);
-
-      authHeaders.delete('content-type');
-      const resp = await fetch('http://localhost:5000/files', {
-        method: 'POST',
-        headers: authHeaders,
-        body: form,
-      });
-      expect(resp.status).toBe(401);
-    } finally {
-      if (fd) {
-        await fd?.close();
-      }
-      if (dir) {
-        await rm(dir, { recursive: true, force: true });
-      }
-    }
+    const { resp } = await dummyFilesSetup(key, fname, projectId, authHeaders);
+    expect(resp.status).toBe(401);
   });
 });
